app/requirements.txt

psutil
requests
python-dotenv
